DataMontgomery Data
This is the scenario of a recently hired Junior Data Analyst in a local government office, who has been tasked with importing some data from another department which relates to inventory information about their fleet of vehicles. I have also been tasked with sorting and analyzing fleet inventory data that was previously imported and cleaned in order to use it to create pivot tables to analyze the data in preparation for the results to be visualized in a dashboard and added to a data findings report later.