Dashboards
This is the scenario of a recently hired Junior Data Analyst in a local government office, who has been tasked with importing some data from another department which relates to inventory information about their fleet of vehicles. The previously imported and cleaned data can now be used to create pivot tables to analyze the data in preparation for the results to be visualized in a dashboard and added to a data findings report later.